// server/src/lib/pokeapi-v2/interfaces.ts
namespace PokeAPI {
  interface APIResource<T> {
    url: APIResourceURL<T>;
  }

  interface NamedAPIResource<T> {
    name: string;
    url: APIResourceURL<T>;
  }
// ...
  export interface Pokemon {
    /** The identifier for this resource. */
    id: number;
    /** The name for this resource. */
    name: string;
    /** The base experience gained for defeating this Pokémon. */
    base_experience: number;
    /** The height of this Pokémon in decimetres. */
    height: number;
    /** Set for exactly one Pokémon used as the default for each species. */
    is_default: boolean;
    /** Order for sorting. Almost national order, except families are grouped together. */
    order: number;
    /** The weight of this Pokémon in hectograms. */
    weight: number;
    /** A list of abilities this Pokémon could potentially have. */
    abilities: PokemonAbility[];
    /** A list of forms this Pokémon can take on. */
    forms: NamedAPIResource<PokemonForm>[];
    /** A list of game indices relevent to Pokémon item by generation. */
    game_indices: VersionGameIndex[];
    /** A list of items this Pokémon may be holding when encountered. */
    held_items: PokemonHeldItem[];
    /** A link to a list of location areas, as well as encounter details pertaining to specific versions. */
    location_area_encounters: string;
    /** A list of moves along with learn methods and level details pertaining to specific version groups. */
    moves: PokemonMove[];
    /** A set of sprites used to depict this Pokémon in the game. A visual representation of the various sprites can be found at <a href='https://github.com/PokeAPI/sprites#sprites'>PokeAPI/sprites</a> */
    sprites: PokemonSprites;
    /** The species this Pokémon belongs to. */
    species: NamedAPIResource<PokemonSpecies>;
    /** A list of base stat values for this Pokémon. */
    stats: PokemonStat[];
    /** A list of details showing types this Pokémon has. */
    types: PokemonType[];
  }

  interface PokemonAbility {
    /** Whether or not this is a hidden ability. */
    is_hidden: boolean;
    /** The slot this ability occupies in this Pokémon species. */
    slot: number;
    /** The ability the Pokémon may have. */
    ability: NamedAPIResource<Ability>;
  }

  interface PokemonType {
    /** The order the Pokémon's types are listed in. */
    slot: number;
    /** The type the referenced Pokémon has. */
    type: NamedAPIResource<Type>;
  }

  interface PokemonHeldItem {
    /** The item the referenced Pokémon holds. */
    item: NamedAPIResource<Item>;
    /** The details of the different versions in which the item is held. */
    version_details: PokemonHeldItemVersion[];
  }

  interface PokemonHeldItemVersion {
    /** The version in which the item is held. */
    version: NamedAPIResource<Version>;
    /** How often the item is held. */
    rarity: number;
  }

  interface PokemonMove {
    /** The move the Pokémon can learn. */
    move: NamedAPIResource<Move>;
    /** The details of the version in which the Pokémon can learn the move. */
    version_group_details: PokemonMoveVersion[];
  }

  interface PokemonMoveVersion {
    /** The method by which the move is learned. */
    move_learn_method: NamedAPIResource<MoveLearnMethod>;
    /** The version group in which the move is learned. */
    version_group: NamedAPIResource<VersionGroup>;
    /** The minimum level to learn the move. */
    level_learned_at: number;
  }

  interface PokemonStat {
    /** The stat the Pokémon has. */
    stat: NamedAPIResource<Stat>;
    /** The effort points (EV) the Pokémon has in the stat. */
    effort: number;
    /** The base value of the stat. */
    base_stat: number;
  }

  interface PokemonSprites {
    /** The default depiction of this Pokémon from the front in battle. */
    front_default: string;
    /** The shiny depiction of this Pokémon from the front in battle. */
    front_shiny: string;
    /** The female depiction of this Pokémon from the front in battle. */
    front_female: string;
    /** The shiny female depiction of this Pokémon from the front in battle. */
    front_shiny_female: string;
    /** The default depiction of this Pokémon from the back in battle. */
    back_default: string;
    /** The shiny depiction of this Pokémon from the back in battle. */
    back_shiny: string;
    /** The female depiction of this Pokémon from the back in battle. */
    back_female: string;
    /** The shiny female depiction of this Pokémon from the back in battle. */
    back_shiny_female: string;
  }
// ...
  interface PokemonForm {
    /** The identifier for this resource. */
    id: number;
    /** The name for this resource. */
    name: string;
    /** The order in which forms should be sorted within all forms. Multiple forms may have equal order, in which case they should fall back on sorting by name. */
    order: number;
    /** The order in which forms should be sorted within a species' forms. */
    form_order: number;
    /** True for exactly one form used as the default for each Pokémon. */
    is_default: boolean;
    /** Whether or not this form can only happen during battle. */
    is_battle_only: boolean;
    /** Whether or not this form requires mega evolution. */
    is_mega: boolean;
    /** The name of this form. */
    form_name: string;
    /** The Pokémon that can take on this form. */
    pokemon: NamedAPIResource<Pokemon>;
    /** A set of sprites used to depict this Pokémon form in the game. */
    sprites: PokemonFormSprites;
    /** The version group this Pokémon form was introduced in. */
    version_group: NamedAPIResource<VersionGroup>;
    /** The form specific full name of this Pokémon form, or empty if the form does not have a specific name. */
    names: Name[];
    /** The form specific form name of this Pokémon form, or empty if the form does not have a specific name. */
    form_names: Name[];
  }
// ...
  interface PokemonSpecies {
    /** The identifier for this resource. */
    id: number;
    /** The name for this resource. */
    name: string;
    /** The order in which species should be sorted. Based on National Dex order, except families are grouped together and sorted by stage. */
    order: number;
    /** The chance of this Pokémon being female, in eighths; or -1 for genderless. */
    gender_rate: number;
    /** The base capture rate; up to 255. The higher the number, the easier the catch. */
    capture_rate: number;
    /** The happiness when caught by a normal Pokéball; up to 255. The higher the number, the happier the Pokémon. */
    base_happiness: number;
    /** Whether or not this is a baby Pokémon. */
    is_baby: boolean;
    /** Whether or not this is a legendary Pokémon. */
    is_legendary: boolean;
    /** Whether or not this is a mythical Pokémon. */
    is_mythical: boolean;
    /** Initial hatch counter: one must walk 255 × (hatch_counter + 1) steps before this Pokémon's egg hatches, unless utilizing bonuses like Flame Body's. */
    hatch_counter: number;
    /** Whether or not this Pokémon has visual gender differences. */
    has_gender_differences: boolean;
    /** Whether or not this Pokémon has multiple forms and can switch between them. */
    forms_switchable: boolean;
    /** The rate at which this Pokémon species gains levels. */
    growth_rate: NamedAPIResource<GrowthRate>;
    /** A list of Pokedexes and the indexes reserved within them for this Pokémon species. */
    pokedex_numbers: PokemonSpeciesDexEntry[];
    /** A list of egg groups this Pokémon species is a member of. */
    egg_groups: NamedAPIResource<EggGroup>[];
    /** The color of this Pokémon for Pokédex search. */
    color: NamedAPIResource<PokemonColor>;
    /** The shape of this Pokémon for Pokédex search. */
    shape: NamedAPIResource<PokemonShape>;
    /** The Pokémon species that evolves into this Pokemon_species. */
    evolves_from_species: NamedAPIResource<PokemonSpecies>;
    /** The evolution chain this Pokémon species is a member of. */
    evolution_chain: APIResource<EvolutionChain>;
    /** The habitat this Pokémon species can be encountered in. */
    habitat: NamedAPIResource<PokemonHabitat> | null;
    /** The generation this Pokémon species was introduced in. */
    generation: NamedAPIResource<Generation>;
    /** The name of this resource listed in different languages. */
    names: Name[];
    /** A list of encounters that can be had with this Pokémon species in pal park. */
    pal_park_encounters: PalParkEncounterArea[];
    /** A list of flavor text entries for this Pokémon species. */
    flavor_text_entries: FlavorText[];
    /** Descriptions of different forms Pokémon take on within the Pokémon species. */
    form_descriptions: Description[];
    /** The genus of this Pokémon species listed in multiple languages. */
    genera: Genus[];
    /** A list of the Pokémon that exist within this Pokémon species. */
    varieties: PokemonSpeciesVariety[];
  }
// ...
  interface VersionGameIndex {
    /** The internal id of an API resource within game data. */
    game_index: number;
    /** The version relevent to this game index. */
    version: NamedAPIResource<Version>;
  }

  interface VersionGroupFlavorText {
    /** The localized name for an API resource in a specific language. */
    text: string;
    /** The language this name is in. */
    language: NamedAPIResource<Language>;
    /** The version group which uses this flavor text. */
    version_group: NamedAPIResource<VersionGroup>;
  }
}